# Conflicts:
#	src/main/java/com/cydeo/repository/CategoryRepository.java
#	src/main/java/com/cydeo/service/CategoryService.java

+++ src/main/java/com/cydeo/service/impl/ClientVendorServiceImpl.java
@@ -7,7 +7,6 @@ import com.cydeo.repository.ClientVendorRepository;
 import com.cydeo.repository.InvoiceRepository;
 import com.cydeo.service.ClientVendorService;
 import com.cydeo.service.CompanyService;
-import com.cydeo.service.SecurityService;
 import com.cydeo.util.MapperUtil;
 import org.springframework.stereotype.Service;
 
@@ -22,14 +21,12 @@ public class ClientVendorServiceImpl implements ClientVendorService {
     private final MapperUtil mapperUtil;
     private final CompanyService companyService;
     private final InvoiceRepository invoiceRepository;
-    private final SecurityService securityService;
 
-    public ClientVendorServiceImpl(ClientVendorRepository clientVendorRepository, MapperUtil mapperUtil, CompanyService companyService, InvoiceRepository invoiceRepository, SecurityService securityService) {
+    public ClientVendorServiceImpl(ClientVendorRepository clientVendorRepository, MapperUtil mapperUtil, CompanyService companyService, InvoiceRepository invoiceRepository) {
         this.clientVendorRepository = clientVendorRepository;
         this.mapperUtil = mapperUtil;
         this.companyService = companyService;
         this.invoiceRepository = invoiceRepository;
-        this.securityService = securityService;
     }
 
     @Override
@@ -49,26 +46,10 @@ public class ClientVendorServiceImpl implements ClientVendorService {
                 .collect(Collectors.toList());
     }
 
-
-    @Override
-    public List<ClientVendorDto> listAllClientVendorsWithInvoiceStatusByCompany() {
-        Long companyId = companyService.getCompanyIdByLoggedInUser();
-        List<ClientVendor> clientVendors = clientVendorRepository.findAllByCompanyIdOrderByTypeAndName(companyId);
-
-        return clientVendors.stream()
-                .map(clientVendor -> {
-                    ClientVendorDto dto = mapperUtil.convert(clientVendor, new ClientVendorDto());
-                    boolean hasInvoice = invoiceRepository.existsByClientVendorId(clientVendor.getId());
-                    dto.setHasInvoice(hasInvoice);
-                    return dto;
-                })
-                .collect(Collectors.toList());
-    }
-
     @Override
     public ClientVendorDto createClientVendor(ClientVendorDto clientVendorDTO) {
 
-        if (existsByName(clientVendorDTO.getClientVendorName())) {
+        if (existsByName(clientVendorDTO.getClientVendorName())){
             throw new IllegalArgumentException("Client/Vendor with this name already exists");
         }
         clientVendorDTO.setCompany(companyService.getCompanyDtoByLoggedInUser());
@@ -80,25 +61,9 @@ public class ClientVendorServiceImpl implements ClientVendorService {
     @Override
     public ClientVendorDto findById(Long id) {
         ClientVendor clientVendor = clientVendorRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("ClientVendor not found"));
+        .orElseThrow(() -> new RuntimeException("ClientVendor not found"));
         return mapperUtil.convert(clientVendor, new ClientVendorDto());
     }
-
-    @Override
-    public List<ClientVendorDto> findAll() {
-        List<ClientVendor> clientVendorList = clientVendorRepository.findAll();
-        return clientVendorList.stream()
-                .map(clientVendor -> mapperUtil.convert(clientVendor, new ClientVendorDto()))
-                .collect(Collectors.toList());
-    }
-//    @Override
-//    public List<ClientVendorDto> findAll() {
-//        List<ClientVendor> clientVendorList = clientVendorRepository.findAll();
-//        return clientVendorList.stream()
-//        .map(clientVendor -> mapperUtil.convert(clientVendor, new ClientVendorDto()))
-//        .collect(Collectors.toList());
-//    }
-
 
     @Override
     public ClientVendorDto updateClientVendor(Long id, ClientVendorDto clientVendorDTO) {
@@ -111,31 +76,23 @@ public class ClientVendorServiceImpl implements ClientVendorService {
     }
 
     @Override
-    public boolean isHasInvoices(Long id) {
-        return invoiceRepository.existsByClientVendorId(id);
-    }
-
     public List<ClientVendorDto> findAllByCurrentCompanyClientVendorTypeAndIsDeleted(ClientVendorType clientVendorType, Boolean isDeleted) {
         Long companyId = companyService.getCompanyIdByLoggedInUser();
-        return mapperUtil.convert(clientVendorRepository.findAllByCompanyIdAndClientVendorTypeAndIsDeleted(companyId, clientVendorType, isDeleted), new ArrayList<>());
+        return mapperUtil.convert(clientVendorRepository.findAllByCompanyIdAndClientVendorTypeAndIsDeleted(companyId,clientVendorType,isDeleted),new ArrayList<>());
     }
-
-//    @Override
-//    public void deleteClientVendor(Long id) {
-//        boolean hasInvoices = invoiceRepository.existsByClientVendorId(id);
-//
-//        if (hasInvoices) {
-//            throw new IllegalStateException("Can not be deleted! This client/vendor has invoice(s).");
-//        }
-//    }
-
     @Override
     public void deleteClientVendor(Long id) {
+        boolean hasInvoices = invoiceRepository.existsByClientVendorId(id);
+
+        if (hasInvoices) {
+            throw new IllegalStateException("Can not be deleted! This client/vendor has invoice(s).");
+        }
+
         clientVendorRepository.deleteById(id);
     }
 
     @Override
-    public boolean existsByName(String clientVendorName){
+    public boolean existsByName(String clientVendorName) {
         return clientVendorRepository.existsByClientVendorName(clientVendorName);
     }
 }
